requirements.txt
requests==2.26.0
pandas==1.5.1
beautifulsoup4==4.11.1
pathlib==1.0.1
xmltodict==0.13.0
pyTelegram==0.0.1
pyTelegramBotAPI==4.7.1
Flask==2.2.2
flask-restx==1.0.3
python-telegram-bot==13.14
schedule==1.1.0
smbprotocol==1.10.1
docxtpl==0.16.5
peewee==3.16.0
gunicorn==20.1.0
peewee_migrate==1.7.1
flask-swagger-ui==4.11.1
emoji==2.2.0
atlassian-python-api==3.39.0
holidays==0.29
graypy==2.1.0
Werkzeug==2.2.2